pyproject.toml
[tool.poetry]
name = "django-rest"
version = "0.1.0"
description = ""
authors = ["Khasanjon Mamadaliyev <khasanjon.eng@gmail.com>"]
readme = "README.md"
packages = [{ include = "root" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
